+++ Assets/Scripts/Stats/Modifier/MultiplicativeModifier.cs
@@ -5,15 +5,14 @@ using Microsoft.CSharp;
 /// <summary>
 /// Generic class for multiplicative modifiers for stats.	
 /// </summary>
-/// <typeparam name="decimal">Any type parameter, which implents the `*` operator</typeparam>
 public class MultiplicativeModifier : Modifier
 {
     public override int Order { get; } = 1000;
 
-    public MultiplicativeModifier(decimal value) : base(value) { }
-    public MultiplicativeModifier(decimal value, StatType statType) : base(value, statType) { }
+    public MultiplicativeModifier(double value) : base(value) { }
+    public MultiplicativeModifier(double value, StatType statType) : base(value, statType) { }
 
-    public override decimal Calculate(decimal value)
+    public override double Calculate(double value)
     {
         return value * (this.value as dynamic);
     }
